+++ test/navbar_utils.dart
@@ -296,7 +296,7 @@ class UserProfile extends StatelessWidget {
             IconButton(
               icon: const Icon(Icons.edit),
               onPressed: () {
-                navigate(context, ProfileEdit.route);
+                navigate(context, ProfileEdit.route, isRootNavigator: false);
               },
             )
           ],